+++ dao/game_result_dao.go
@@ -7,6 +7,6 @@ import (
 )
 
 type GameResultDAO interface {
-	CreateGameResult(ctx context.Context, userId uuid.UUID, gameStatus entity.GameStatus, Amount float64, transactionID string) (*entity.GameResult, error)
+	CreateGameResult(ctx context.Context, userId uuid.UUID, gameStatus entity.GameStatus, amount float64, transactionSource entity.TransactionSource, transactionID string) (*entity.GameResult, error)
 	ValidateGameResults(ctx context.Context, totalGamesToCancel int) error
 }