+++ dao/game_result_dao_implementation.go
@@ -1,0 +1,109 @@
+package dao
+
+import (
+	"context"
+	"github.com/google/uuid"
+	"github.com/ildomm/cceab/database"
+	"github.com/ildomm/cceab/entity"
+	"github.com/jmoiron/sqlx"
+	"log"
+)
+
+type gameResultDAO struct {
+	querier database.Querier
+}
+
+func NewGameResultDAO(querier database.Querier) *gameResultDAO {
+	dm := gameResultDAO{
+		querier: querier,
+	}
+	return &dm
+}
+
+// CreateGameResult creates a new game result in the database.
+func (dm *gameResultDAO) CreateGameResult(ctx context.Context, userId uuid.UUID, gameStatus entity.GameStatus, amount float64, transactionSource entity.TransactionSource, transactionID string) (*entity.GameResult, error) {
+
+	// Validate the existence of the transaction ID.
+	exists, err := dm.querier.CheckTransactionID(ctx, transactionID)
+	if err != nil {
+		return nil, err
+	}
+
+	if exists {
+		return nil, entity.ErrTransactionIdExists
+	}
+
+	// Validate the user.
+	user, err := dm.querier.SelectUser(ctx, userId)
+	if err != nil {
+		return nil, err
+	}
+
+	if user == nil {
+		return nil, entity.ErrUserNotFound
+	}
+
+	// Check if the user has a negative balance.
+	if gameStatus == entity.GameStatusLost && user.Balance < amount {
+		return nil, entity.ErrUserNegativeBalance
+	}
+
+	// Define the new balance.
+	balance := user.Balance
+	if gameStatus == entity.GameStatusWin {
+		balance = user.Balance + amount
+	} else {
+		balance = user.Balance - amount
+	}
+
+	// Define the game result.
+	gameResult := entity.GameResult{
+		UserID:            userId,
+		GameStatus:        gameStatus,
+		ValidationStatus:  entity.ValidationStatusPending,
+		TransactionSource: transactionSource,
+		TransactionID:     transactionID,
+		Amount:            amount,
+	}
+
+	// Starts a new transaction
+	err = dm.querier.WithTransaction(ctx, func(txn *sqlx.Tx) error {
+
+		// Lock the user row
+		err := dm.querier.LockUserRow(ctx, *txn, userId)
+		if err != nil {
+			return err
+		}
+
+		// Create the game result
+		id, err := dm.querier.InsertGameResult(ctx, *txn, gameResult)
+		if err != nil {
+			return err
+		}
+		gameResult.ID = id
+
+		// Update the user balance
+		err = dm.querier.UpdateUserBalance(ctx, *txn, userId, balance, false)
+		if err != nil {
+			return err
+		}
+
+		// Commit the transaction
+		// Success, continue with the transaction commit
+		return nil
+	})
+
+	// In case of error in the transactional operation, log and exit
+	if err != nil {
+		log.Printf("error performing game result db transaction: %v", err)
+		return nil, entity.ErrCreatingGameResult
+	}
+
+	return nil, nil
+}
+
+// ValidateGameResults validates the game results in post-game processing.
+func (dm *gameResultDAO) ValidateGameResults(ctx context.Context, totalGamesToCancel int) error {
+
+	return nil
+}