+++ server/middlewares_test.go
@@ -1,7 +1,6 @@
 package server
 
 import (
-	"github.com/ildomm/cceab/entity"
 	"github.com/ildomm/cceab/test_helpers"
 	"net/http"
 	"net/http/httptest"
@@ -66,36 +65,4 @@ func TestLoggingMiddleware(t *testing.T) {
 	assert.Contains(t, logOutput, "INFO", "log does not contain info level")
 	assert.Contains(t, logOutput, "202", "log does not contain correct status code")
 	assert.Contains(t, logOutput, "ms", "log does not contain execution time")
-}
-
-// TestSourceTypeValidatorMiddleware tests the SourceTypeValidatorMiddleware's handling of valid and invalid source types.
-func TestSourceTypeValidatorMiddleware(t *testing.T) {
-	validHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusOK)
-	})
-	sourceTypeValidatorMiddleware := NewSourceTypeValidatorMiddleware()
-
-	// Test with valid source type
-	testServerValid := httptest.NewServer(sourceTypeValidatorMiddleware(validHandler))
-	defer testServerValid.Close()
-
-	reqValid, err := http.NewRequest("GET", testServerValid.URL, nil)
-	assert.NoError(t, err)
-	reqValid.Header.Set("Source-Type", string(entity.TransactionSourceGame))
-
-	respValid, err := http.DefaultClient.Do(reqValid)
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusOK, respValid.StatusCode, "middleware incorrectly handled valid source type")
-
-	// Test with invalid source type
-	testServerInvalid := httptest.NewServer(sourceTypeValidatorMiddleware(validHandler))
-	defer testServerInvalid.Close()
-
-	reqInvalid, err := http.NewRequest("GET", testServerInvalid.URL, nil)
-	assert.NoError(t, err)
-	reqInvalid.Header.Set("Source-Type", "invalid")
-
-	respInvalid, err := http.DefaultClient.Do(reqInvalid)
-	assert.NoError(t, err)
-	assert.Equal(t, http.StatusBadRequest, respInvalid.StatusCode, "middleware did not handle invalid source type correctly")
 }